refactor(bigquery): move in write client
