+++ .cargo/config.toml
@@ -22,8 +22,6 @@ clippy-strict = [
   "-p",
   "google-cloud-bigquery",
   "-p",
-  "google-cloud-bigquery-write",
-  "-p",
   "google-cloud-bigtable",
   "-p",
   "google-cloud-datastore",
