+++ Cargo.toml
@@ -21,7 +21,6 @@ default-members = [
   "src/bigquery",
   "src/bigquery-derive",
   "src/bigquery-read",
-  "src/bigquery-write",
   "src/bigquery/examples",
   "src/bigtable",
   "src/datastore",
@@ -71,7 +70,6 @@ members = [
   "src/bigquery",
   "src/bigquery-derive",
   "src/bigquery-read",
-  "src/bigquery-write",
   "src/bigquery-write/grpc-mock",
   "src/bigquery/examples",
   "src/bigtable",
@@ -582,7 +580,6 @@ google-cloud-bigquery                = { default-features = false, path = "src/b
 google-cloud-bigquery-derive         = { default-features = false, path = "src/bigquery-derive", version = "0.1.0" }
 google-cloud-bigquery-read           = { default-features = false, path = "src/bigquery-read" }
 google-cloud-bigquery-v2             = { default-features = false, path = "src/generated/cloud/bigquery/v2", version = "0.1.0" }
-google-cloud-bigquery-write          = { default-features = false, path = "src/bigquery-write", version = "0.1.0" }
 google-cloud-iam-credentials-v1      = { default-features = false, path = "src/generated/iam/credentials/v1" }
 google-cloud-language-v2             = { default-features = false, path = "src/generated/cloud/language/v2" }
 google-cloud-dns-v1                  = { default-features = false, path = "src/generated/cloud/dns/v1" }
