--- src/bigquery-write/Cargo.toml
@@ -1,63 +1,0 @@
-# Copyright 2026 Google LLC
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     https://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-[package]
-name        = "google-cloud-bigquery-write"
-version     = "0.1.0"
-description = "Google Cloud Client Libraries for Rust - BigQuery Write"
-# Inherit other attributes from the workspace.
-edition.workspace      = true
-authors.workspace      = true
-license.workspace      = true
-repository.workspace   = true
-keywords.workspace     = true
-categories.workspace   = true
-rust-version.workspace = true
-
-[features]
-default = ["default-rustls-provider"]
-# Enabled by default. Use the default rustls crypto provider ([aws-lc-rs]) for
-# TLS and authentication. Applications with specific requirements for
-# cryptography (such as exclusively using the [ring] crate) should disable this
-# default and call `rustls::CryptoProvider::install_default()`.
-default-rustls-provider = ["gaxi/_default-rustls-provider"]
-
-[dependencies]
-async-trait.workspace       = true
-bytes.workspace             = true
-http.workspace              = true
-prost.workspace             = true
-prost-types.workspace       = true
-serde.workspace             = true
-serde_json.workspace        = true
-serde_with.workspace        = true
-thiserror.workspace         = true
-tokio.workspace             = true
-tokio-stream.workspace      = true
-tracing.workspace           = true
-google-cloud-auth.workspace = true
-google-cloud-gax.workspace  = true
-google-cloud-rpc.workspace  = true
-wkt.workspace               = true
-gaxi                        = { workspace = true, features = ["_internal-common", "_internal-grpc-client"] }
-
-[dev-dependencies]
-anyhow.workspace            = true
-test-case.workspace         = true
-bigquery-write-grpc-mock    = { path = "grpc-mock" }
-google-cloud-auth.workspace = true
-google-cloud-bigquery-write = { path = ".", features = ["default-rustls-provider"] }
-
-[lints]
-workspace = true
