--- src/bigquery-write/README.md
@@ -1,10 +1,0 @@
-# Google Cloud Client Libraries for Rust - BigQuery Write
-
-This crate implements the [BigQuery Write] client library.
-
-**WARNING:** this crate is under active development. We expect multiple breaking
-changes in the upcoming releases. Testing is also incomplete, we do **not**
-recommend that you use this crate in production. We welcome feedback about the
-APIs, documentation, missing features, bugs, etc.
-
-[bigquery write]: https://docs.cloud.google.com/bigquery/docs/write-api
