+++ src/bigquery/src/write/append_response.rs
@@ -12,10 +12,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+use super::error::{AppendError, AppendResult};
+use super::generated::gapic_storage::model::append_rows_response::Response;
+use super::model::{AppendRowsResponse, TableSchema};
 use crate::Error;
-use crate::error::{AppendError, AppendResult};
-use crate::generated::gapic_storage::model::append_rows_response::Response;
-use crate::model::{AppendRowsResponse, TableSchema};
 
 /// The return type of an `append()` operation.
 #[derive(Clone, Debug, Default, PartialEq)]
@@ -57,9 +57,9 @@ pub(crate) fn to_result(resp: AppendRowsResponse) -> AppendResult<AppendResponse
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::model::RowError;
-    use crate::model::append_rows_response::AppendResult;
-    use crate::model::row_error::RowErrorCode;
+    use crate::write::model::RowError;
+    use crate::write::model::append_rows_response::AppendResult;
+    use crate::write::model::row_error::RowErrorCode;
     use google_cloud_gax::error::rpc::Code;
     use google_cloud_rpc::model::Status as RpcStatus;
