+++ src/bigquery/src/write/arrow/default.rs
@@ -12,41 +12,37 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use crate::Result;
-use crate::builder::write::AppendWithOffset;
-use crate::generated::gapic_storage::client::BigQueryWrite;
-use crate::model::append_rows_request::ArrowData;
-use crate::model::{
-    AppendRowsRequest, ArrowRecordBatch, ArrowSchema, BatchCommitWriteStreamsResponse,
-    FinalizeWriteStreamResponse,
-};
-use crate::runner::Runner;
-use crate::transport::Transport;
+use super::super::append_builder::Append;
+use super::super::model::append_rows_request::ArrowData;
+use super::super::model::{AppendRowsRequest, ArrowRecordBatch, ArrowSchema};
+use super::super::runner::Runner;
+use super::super::transport::Transport;
 use std::sync::Arc;
 
-/// A writer for a pending stream.
+/// A writer for the [default stream]
+///
+/// [default stream]: https://docs.cloud.google.com/bigquery/docs/write-api#default_stream
 #[derive(Debug)]
-pub struct PendingWriter {
+pub struct DefaultWriter {
+    // TODO(#5744) - support multiplexed connections
     runner: Runner,
     pub(crate) write_stream: String,
     pub(crate) schema: ArrowSchema,
-    client: BigQueryWrite,
 }
 
-impl PendingWriter {
+impl DefaultWriter {
     pub(crate) fn new(inner: Arc<Transport>, write_stream: String, schema: ArrowSchema) -> Self {
-        let runner = Runner::new(inner.clone());
-        let client = BigQueryWrite::from_stub::<Transport>(inner);
+        let runner = Runner::new(inner);
         Self {
             runner,
             write_stream,
             schema,
-            client,
         }
     }
 
-    /// Appends rows to the pending stream.
-    pub fn append(&self, rows: ArrowRecordBatch) -> AppendWithOffset {
+    /// Append rows to the stream.
+    pub fn append(&self, rows: ArrowRecordBatch) -> Append {
+        // TODO(#5744) - send optimization
         let req = AppendRowsRequest::new()
             .set_write_stream(&self.write_stream)
             .set_arrow_rows(
@@ -54,42 +50,16 @@ impl PendingWriter {
                     .set_writer_schema(self.schema.clone())
                     .set_rows(rows),
             );
-        AppendWithOffset::new(self.runner.req_tx.clone(), req)
-    }
-
-    /// Finalizes the pending stream, preventing further writes.
-    pub async fn finalize(&self) -> Result<FinalizeWriteStreamResponse> {
-        self.client
-            .finalize_write_stream()
-            .set_name(&self.write_stream)
-            .send()
-            .await
-    }
-
-    /// Commits the pending stream to the table.
-    pub async fn commit(&self) -> Result<BatchCommitWriteStreamsResponse> {
-        // Extract the parent table path from the stream name:
-        // "projects/p/datasets/d/tables/t/streams/s" -> "projects/p/datasets/d/tables/t"
-        let parent = self
-            .write_stream
-            .split_once("/streams/")
-            .map_or(self.write_stream.as_str(), |(p, _)| p)
-            .to_string();
-
-        self.client
-            .batch_commit_write_streams()
-            .set_parent(parent)
-            .set_write_streams(vec![self.write_stream.clone()])
-            .send()
-            .await
+        Append::new(self.runner.req_tx.clone(), req)
     }
 }
 
 #[cfg(test)]
 mod tests {
+    use super::super::super::runner::tests::*;
+    use super::super::super::transport::tests::*;
     use super::*;
-    use crate::runner::tests::*;
-    use crate::transport::tests::*;
+    use crate::error::AppendError;
     use bigquery_write_grpc_mock::{MockBigQueryWrite, start};
     use gaxi::grpc::tonic::Response as TonicResponse;
     use tokio::sync::mpsc;
@@ -97,7 +67,7 @@ mod tests {
     #[tokio::test]
     async fn request_fields() -> anyhow::Result<()> {
         let transport = Arc::new(test_transport("http://ignored:1".to_string()).await?);
-        let writer = PendingWriter::new(transport, write_stream(), schema());
+        let writer = DefaultWriter::new(transport, write_stream(), schema());
 
         let b = writer.append(rows(1));
         assert_eq!(b.req.write_stream, write_stream());
@@ -106,6 +76,14 @@ mod tests {
         assert_eq!(s.serialized_schema, "test");
         let r = data.rows.as_ref().expect("rows should be set");
         assert_eq!(r.serialized_record_batch, "1");
+
+        let b = writer.append(rows(2));
+        assert_eq!(b.req.write_stream, write_stream());
+        let data = b.req.arrow_rows().expect("arrow rows should be set");
+        let s = data.writer_schema.as_ref().expect("schema should be set");
+        assert_eq!(s.serialized_schema, "test");
+        let r = data.rows.as_ref().expect("rows should be set");
+        assert_eq!(r.serialized_record_batch, "2");
 
         Ok(())
     }
@@ -117,34 +95,32 @@ mod tests {
         let mut mock = MockBigQueryWrite::new();
         mock.expect_append_rows()
             .return_once(|_| Ok(TonicResponse::from(response_rx)));
-
-        mock.expect_finalize_write_stream()
-            .return_once(|_| Ok(TonicResponse::new(
-                bigquery_write_grpc_mock::google::cloud::bigquery::storage::v1::FinalizeWriteStreamResponse::default()
-            )));
-
-        mock.expect_batch_commit_write_streams()
-            .return_once(|_| Ok(TonicResponse::new(
-                bigquery_write_grpc_mock::google::cloud::bigquery::storage::v1::BatchCommitWriteStreamsResponse::default()
-            )));
-
         let (endpoint, _server) = start("0.0.0.0:0", mock).await?;
         let transport = Arc::new(test_transport(endpoint).await?);
 
-        let writer = PendingWriter::new(transport, write_stream(), schema());
+        let writer = DefaultWriter::new(transport, write_stream(), schema());
 
         response_tx.send(Ok(convert(&test_response(1)))).await?;
         let resp = writer.append(rows(1)).send().await?;
         assert_eq!(resp.offset, Some(1));
 
-        writer.finalize().await?;
-        writer.commit().await?;
+        response_tx.send(Ok(convert(&test_response(2)))).await?;
+        let resp = writer.append(rows(2)).send().await?;
+        assert_eq!(resp.offset, Some(2));
+
+        response_tx.send(Ok(convert(&test_response(3)))).await?;
+        let resp = writer.append(rows(3)).send().await?;
+        assert_eq!(resp.offset, Some(3));
+
+        drop(response_tx);
+        let err = writer.append(rows(4)).send().await.expect_err("channel");
+        assert!(matches!(err, AppendError::UnexpectedEndOfStream));
 
         Ok(())
     }
 
     fn write_stream() -> String {
-        "projects/p/datasets/d/tables/t/streams/s".to_string()
+        "projects/p/datasets/d/tables/t/streams/_default".to_string()
     }
 
     fn schema() -> ArrowSchema {
