+++ src/bigquery/src/write/generated.rs
